Move the session log to common and its export to data
The session log is written from data/, portfolio/, and handlers.rs, so it
belongs in the module every layer already depends on rather than in one of
them. Moving it there needed SessionDate to move too: common/ imports nothing
from data/ today, and leaving SessionDate behind would have inverted that for
the sake of one type. It is a domain primitive in the same sense Ticker and
Price are, so common/types.rs is where it should have been. TradingCalendar
stays in data/calendar.rs, because it really does fetch from Alpaca.

The export half did not follow. Sealing a file, converting it to Parquet, and
shipping it to S3 is called from exactly one place, handle_database_export,
next to the database export doing the same job — and the two had grown
identical write_frame functions. Folding the session log export into
data/export.rs removes that duplicate and leaves common/session_log.rs as the
record types and the writer, which is the part that is actually cross-cutting.

Sealing crosses the new boundary, so the guard is now a named opaque type
rather than a bare MutexGuard over a private struct. Holding it is the whole
contract; there is nothing inside worth exposing.

No behavior change. Tests moved with the code they cover.
