+++ src/portfolio/evaluate.rs
@@ -17,13 +17,13 @@ use tokio_util::sync::CancellationToken;
 use tracing::{debug, error, info, warn};
 
 use crate::common::alpaca::{ClientError, MarketDataClient, PriceSource, TradingClient};
-use crate::common::types::{EquityPrediction, Ticker};
-use crate::data::calendar::{SessionDate, TradingCalendar};
-use crate::data::details::{self, DetailsError};
-use crate::data::session_log::{
+use crate::common::session_log::{
     CandidateReading, EvaluationPass, ExcludedTickerReading, LiquidationRun, Observation,
     OpenPairReading, PositionCloseRequested, PriceReading, ScreenInputReading, SessionLog,
 };
+use crate::common::types::{EquityPrediction, SessionDate, Ticker};
+use crate::data::calendar::TradingCalendar;
+use crate::data::details::{self, DetailsError};
 use crate::data::universe::Universe;
 use crate::models::tide::predict;
 use crate::portfolio::account::{self, AccountError};
